more MIR const types to separate file
